+++ changelog.md
@@ -1,3 +1,11 @@
+Patch 3689 (30th September 2017)
+============================
+### Bugs
+- Veterancy
+    - Fixed a bug that caused units to give 100% veterancy points instead of a fraction depending on their remaining HP on death when getting destroyed by self destruction.
+    - Fixed ACUs giving too much veterancy to other ACUs when being killed.
+- Fixed shields recharging many times faster than intended.
+
 Patch 3688 (24th September 2017)
 ============================
 ### Balance
@@ -9,6 +17,14 @@ Patch 3688 (24th September 2017)
     - Experimental units only need to kill 50% of their own worth in mass to gain one veterancy rank.
     - Regular units need to destroy 200% of their own worth in mass to gain one veterancy rank.
     - Destroying a transport gives veterancy points for the cargo as well as the transport itself.
+    - Unified regeneration values.
+        - T1      +1  ->  1,  2,  3,   4,   5 regeneration
+        - T2      +3  ->  3,  6,  9,  12,  15 regeneration
+        - T3/ACUs +6  ->  6, 12, 18,  24,  30 regeneration
+        - SACUs   +9  ->  9, 18, 27,  36,  45 regeneration
+        - T4      +25 -> 25, 50, 75, 100, 125 regeneration
+- Shields
+    - Reduced the time a shield needs before it starts to recharge.
 - Gunships
     - Seraphim T2 Gunship
         - Health: 864 → 1080